updated interaction model to reflect use scenario
I updated the AI Adaptive LD System Interactions Diagram to match the use scenario and also made a few edits to the rest of the page to make it consistent. I also added the links to other diagram that it may connect to.

questions

Should there be an arrow to course designer to cloud depository to represent that the CD can also add data or is that redundant as the CD may have the function to deposit data within the PE itself? Regarding how users deposit to the CCD, is this a separate interface?