Changed the Interactions Diagrams for LAP
I changed the diagram to only fit the Use Case Scenario model because when I compared it to the Parent Interaction Diagram it didn't follow the same logic. In the parent interaction diagram the Learning Analytics proposes resources and searches the web for resources and I don't think that's something the LAP should do. I think yesterday we agreed that the LAP will only feed data to the Personalisation Engine and Cloud Repository and in such a case the PE will interact with the web to search for resources. 
We need to change the *Parent Interactions Model* to reflect the changes we agreed upon yesterday in terms of the LAP behaviour that Daniela suggested.