Added Diagram
Added Quang's Adaptive LD diagram. I feel it helps to demonstrate the description above. I know there was feedback that it would be too much for the README however, this page is an explanatory page and I do feel it is beneficial to have visual representation. Happy to remove if no one else agrees.