attempted to fix the link
I know this is a minor change so apologies for the pull request but I have had this issue multiple times where the image appears as it should on my screen but for others it doesn't appear. I'm still not 100% sure why that is (anyone?) but I wanted to see if this works before I commit another broken link to the master.