Course Design with Smart Functonality
Thought I would put this up for discussion. I ended up doing a use case without AI as an actor because it didn't need to be.

What if the entire LE has smart functionality, as stored user data and made decisions within the platform without an other actors interfering?

What if a CD and COP only added to the repository? So was able to elevate the experience by providing more resources for AI to read from? I think this is a good start but could be expanded upon. Especially if Daniela has her own ideas coming out of the research she has done.

Let me know of your thoughts.